Move the exceptions interface to its own module.
